Remove Oj as a required dependency
This gem uses multi_json under the hood to choose the faster JSON coder
available in the application.  By requiring Oj as a dependency, we
remove that choice and require the application to deserialize and
re-serialize each response using Oj, which may result in a different
JSON response.  For example, Oj sometimes de-serializes floating numbers
as BigDecimals instead of Floats.  BigDecimals then get serialized as
strings instead of the original floating number.  This is a different
implementation from the JSON coder that ships with Ruby.

As this is is a breaking change from 3.0.0, this should be considered a
new major version.